feat(communication_layer): Implemented return for serialize function
The `DUUILuaCompiledFile.call()` method now returns a `LuaTable` that is
passed back to the communication layer class.
If the return is `nil` an empty table is returned (ensures backwards
compatability).
If a value of another type is returned, a `LuaError` is thrown.

The different `call(...)` implementations were unified into a single
method that takes any number of arguments.

`IDUUICommunicationLayer` now returns a `SerializeOutput` for
`serialize` calls. This object is constructed from the `LuaTable`
returned by the function `call()` and currently only contains a field
for request headers to be passed to the component.

The `deserialize` method has not been changed.