fix(LuaJ): Fixed LuaJ casting all numbers to JSON floats
LuaJ's implementation of the Lua `string.format` function causes the
formatting string `%.14g` used by json.lua to format numbers to always
append a `.0` for integer values. This differs from the default Lua
implementation (v5.4). As such, applications written in typesafe
languages like Go or Rust can NOT be addressed properly using JSON
serialized data.

I fixed this by checking if the number is actually an integer and
chaning the formatting string to `%d` if it is.

This should not have negative effects downstream as JSON deserialization
can easily coerce integers into floats, but not the otherway around.